# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a backend server for service mini.note resource.",
        "title": "Mini Note RESTful API",
        "contact": {
            "name": "API Support"
        },
        "version": "0.1"
    },
    "host": "localhost:9077",
    "basePath": "/api/v1",
    "paths": {
...
        "/send/request": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send message to email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Request"
                ],
                "summary": "Send request",
                "parameters": [
                    {
                        "description": "Request param",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/UniversalDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1
                },
                "message": {
                    "type": "string",
                    "example": "Note not found."
                }
            }
        },
...
        "Request": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Jhon Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "+79000000000"
                }
            }
        },
        "UniversalDTO": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/Error"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}